rename checkToken to spotifyAuth
